pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.55.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
